Make DSEC ROS support optional
Move DSEC ROS bag reader and decoder helpers into a private module.
